==> Scripts/05_Differential_Expression.Rmd <==
---
title: "Differential expression and marker identification"
author: "Konstantinos Charitidis"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
output:
  html_document:
    toc: true
    toc_depth: '3'
    df_print: paged
  bookdown::html_document2:
    toc: true
    toc_depth: 3
    toc_float:
      collapse: false
    number_sections: true
    df_print: paged
    theme: readable
    code_folding: hide
    self_contained: true
    highlight: tango
---

```{r Setup, include = FALSE}

knitr::opts_chunk$set(echo = TRUE, eval = TRUE, message = FALSE, warning = FALSE)
options(max.print = 200)

```

```{r Libraries, include = FALSE}

library(tidyverse)
library(Seurat)
library(SeuratObject)
library(Matrix)
library(RColorBrewer)
library(ggtext)
library(scales)
library(patchwork)
source(file = "~/P_lab/Spatial_CART/Scripts/Other/My_Functions.R")
Init_project("~/P_lab/Spatial_CART/")

```

```{r Load the batch corrected seurat oject}

Seurat_obj <- readRDS(file = file.path(main_dir, "Data/Edited_data/Seurat_Annotated.Rds"))

```

```{r Highly variable features}

HV_Genes <- HVFInfo(Seurat_obj) %>%
  as.data.frame() %>%
  tibble::rownames_to_column("Gene")

top_20_HVF <- head(VariableFeatures(Seurat_obj), 20)
total_HVF <- VariableFeatures(Seurat_obj)

HV_Genes <- HV_Genes %>% 
  mutate(is_top = Gene %in% top_20_HVF,
         is_HVF = Gene %in% total_HVF,
         Label = ifelse(is_top, Gene, NA_character_),
         Rank = rank(-variance.standardized, ties.method = "first"), 
         Var_class = if_else(is_HVF, "Variable", "Non-variable"))

n_var <- sum(HV_Genes$Var_class == "Variable", na.rm = TRUE)
n_non <- sum(HV_Genes$Var_class == "Non-variable", na.rm = TRUE)

HVF_plot <- ggplot(HV_Genes, aes(x = mean, y = variance.standardized)) +
  geom_point(aes(alpha = is_top, size = 0.1, colour = Var_class)) +
  ggrepel::geom_label_repel(data = HV_Genes %>% dplyr::filter(is_top),
                           aes(label = Label), size = 3, max.overlaps = Inf,
                           box.padding = 0.4, point.padding = 0.2, 
                           fill = "white", colour = "black") +
  scale_color_manual(values = c("Non-variable" = "grey30", "Variable" = "firebrick"), 
                     breaks = c("Non-variable", "Variable"),
                     labels = c("Non-variable" = paste0("Non-variable count: ", n_non),
                                "Variable" = paste0("Variable count: ", n_var)),
                     name = NULL) +
  scale_alpha_manual(values = c(`FALSE` = 0.8, `TRUE` = 1)) +
  scale_x_log10() +
  scale_y_log10() +
  labs(x = "Average expression (log<sub>10</sub>)",
       y = "Standardized variance (log<sub>10</sub>)",
       title = paste0("Highly variable genes (top 20 highlighted)")) +
  theme_minimal(base_size = 12) +
  theme(plot.title = element_text(face = "bold"),
        axis.title.x = element_markdown(face = "bold"),
        axis.title.y = element_markdown(face = "bold"),
        panel.grid.minor = element_blank()) +
  guides(alpha = "none", size = "none")

# ggsave(filename = file.path(main_dir, "Results/DE_Plots/HVG_rank_plot.png"),
#        plot = HVF_plot, width = 16, height = 10, dpi = 600)

```

```{r Elbow plot}

# Get standard deviations from PCA
PCA_sdev <- Seurat_obj[["pca"]]@stdev

Elbow_df <- tibble(PC = 1:length(PCA_sdev),
                   Variance = PCA_sdev^2) %>%
  mutate(PropVar = Variance / sum(Variance), 
         CumPropVar = cumsum(PropVar))

dims_to_highlight <- min(Elbow_df$PC[Elbow_df$CumPropVar >= 0.6])

Elbow_plot <- ggplot(Elbow_df, aes(x = PC, y = PropVar)) +
  geom_line() +
  geom_point(size = 1) +
  geom_vline(xintercept = dims_to_highlight, linetype = "dashed") +
  annotate("text", x = dims_to_highlight, y = max(Elbow_df$PropVar) * 0.5,
           label = paste0("60% variance at PC ", dims_to_highlight),
           angle = 90, size = 3, vjust = -0.5) +
  scale_y_continuous(labels = percent_format(accuracy = 1), 
                     limits = c(0, max(Elbow_df$PropVar) * 1.05), expand = c(0, 0)) +
  scale_x_continuous(limits = c(1, 20), breaks = seq(1, 20, by = 1)) +
  labs(x = "Principal component",
       y = "Variance explained",
       title = "Elbow plot for PCA") +
  theme_classic(base_size = 12) +
  theme(plot.title = element_text(face = "bold"),
        axis.title.x = element_text(face = "bold"),
        axis.title.y = element_text(face = "bold"), 
        panel.grid.minor = element_blank(), 
        panel.grid.major = element_blank()) 

Elbow_plot


```

```{r VizDimLoadings improved}

dims_to_plot <- 1:12

PCA_loadings <- Seurat_obj[["pca"]]@feature.loadings

Loadings_df <- as.data.frame(PCA_loadings) %>%
  tibble::rownames_to_column("Gene") %>%
  dplyr::select(Gene, dplyr::all_of(paste0("PC_", dims_to_plot))) %>%
  tidyr::pivot_longer(cols = dplyr::starts_with("PC_"),
                      names_to = "PC",
                      values_to = "Loading") %>%
  dplyr::group_by(PC) %>%
  dplyr::mutate(abs_loading = abs(Loading)) %>%
  dplyr::slice_max(n = 20, order_by = abs_loading) %>%
  dplyr::ungroup() %>%
  dplyr::group_by(PC) %>%
  dplyr::mutate(Gene = forcats::fct_reorder(Gene, Loading), 
                PC = factor(PC, levels = paste0("PC_", dims_to_plot), labels = paste("PC", dims_to_plot))) %>%
  dplyr::ungroup()

PC_Loadings_plot <- ggplot(Loadings_df, aes(x = Gene, y = Loading, fill = Loading)) +
  geom_col() +
  coord_flip() +
  scale_fill_gradient2(low = "royalblue", mid = "white", high = "firebrick") +
  facet_wrap(~ PC, scales = "free_y") +
  labs(x = NULL, y = "PC loading",
       title = "Top 20 genes by PC loading",
       fill = "Loading") +
  theme_minimal(base_size = 12) +
  theme(plot.title  = element_text(face = "bold"),
        axis.title.x = element_text(face = "bold"),
        # axis.title.y = element_text(face = "bold"),
        panel.grid.minor = element_blank(),
        panel.grid.major = element_blank())

PC_Loadings_plot

```

```{r}

# print(Seurat_obj[["pca"]], dims = 1:5, nfeatures = 5)
VizDimLoadings(Seurat_obj, dims = 1:12, reduction = "pca")
DimHeatmap(Seurat_obj, dims = 1:12, cells = 1500, balanced = TRUE)
ElbowPlot(Seurat_obj)

```

```{r Compute DE markers per cluster}

markers_Kidney_HCA <- FindAllMarkers(Seurat_obj,
                                     assay = "RNA", 
                                     group.by = "Kidney_HCA_clusters",
                                     only.pos = TRUE, 
                                     min.pct = 0.05, ## gene expressed in >= 5% of cells 
                                     logfc.threshold = 0.5,
                                     test.use = "wilcox")

markers_unsupervised <- FindAllMarkers(Seurat_obj,
                                       assay = "RNA", 
                                       group.by = "Unsupervised_3_clusters",
                                       only.pos = TRUE, 
                                       min.pct = 0.05, ## gene expressed in >= 5% of cells 
                                       logfc.threshold = 0.5,
                                       test.use = "wilcox")

# write_csv(x = markers_Kidney_HCA, file = file.path(main_dir, "Results/Tables/DE_Genes/markers_Kidney_HCA.csv"))
# write_csv(x = markers_unsupervised, file = file.path(main_dir, "Results/Tables/DE_Genes/markers_unsupervised_3.csv"))

```

```{r Violin plots for the Kidney HCA dataset}

Custom_VLN(obj = Seurat_obj, 
           features = c("ALDOB"),
           group_by = "Kidney_HCA_clusters",
           ncol = 1)


plots_HCA <- Plot_top_markers_VLN(obj = Seurat_obj,
                                  markers_tbl = markers_Kidney_HCA,
                                  group_by = "Kidney_HCA_clusters",
                                  out_dir = file.path(main_dir, "Results/DE_Plots/VLN_Plots/Kidney_HCA"),
                                  top_n = 3,
                                  prefix = "Kidney_HCA")

B_pan <- c("MS4A1", "CD79A", "CD79B", "CD19", "CD74", "CD22")
intersect(B_pan, Features(Seurat_obj))
gene_tag <- paste(B_pan, collapse = "_")

p1 <- Plot_manual_markers_VLN(obj = Seurat_obj,
                              group_by = "Kidney_HCA_clusters",
                              cell_type = "B.cell",
                              markers = B_pan,
                              save_path = file.path(main_dir, "Results/DE_Plots/VLN_Plots/Custom_Kidney_HCA",
                              paste0("B_cell_", gene_tag, ".png")),
                              pt_size = 0)

```

```{r B cell marker plots}

B_top_3 <- markers_Kidney_HCA %>% 
  filter(cluster == "B.cell") %>% 
  slice_max(n = 3, order_by = avg_log2FC) %>% 
  pull(gene)

B_main <- c("MS4A1", "CD19", "CD22")
B_Markers <- c(B_top_3, B_main)

p_MS4A1 <- FeaturePlot(Seurat_obj, 
                       features = B_Markers,
                       reduction = "umap",   
                       pt.size = 1,
                       min.cutoff = "q05",
                       max.cutoff = "q95", 
                       # cols = c("grey90", "firebrick")
                       combine = FALSE) +
  theme_void(base_size = 12) +
  theme(plot.title = element_text(face = "bold", hjust = 0.5),
        legend.position = "right",
        legend.title = element_text(face = "bold"),
        legend.text = element_text(size = 9)) +
  labs(title = "B cell markers expression on the UMAP projection", colour = "Expression")

p_MS4A1

```

```{r B cell markers feature plots}

B_plots <- FeaturePlot(
  Seurat_obj,
  features   = B_Markers,
  reduction  = "umap",
  pt.size    = 0.4,
  min.cutoff = "q05",
  max.cutoff = "q95",
  cols       = c("grey90", "midnightblue"),
  combine    = FALSE           # <- important
)

# 2. Clean theme for each plot
B_plots <- lapply(seq_along(B_plots), function(i) {
  B_plots[[i]] +
    theme_void(base_size = 12) +
    theme(
      plot.title   = element_text(face = "bold", hjust = 0.5),
      legend.position = "right",
      legend.title    = element_text(face = "bold"),
      legend.text     = element_text(size = 8),
      plot.margin     = margin(t = 5, r = 5, b = 5, l = 5)
    ) +
    labs(title = B_Markers[i], colour = "Expression")
})

# 3. Arrange in a grid and add a single global title + single legend
B_feature_grid <- wrap_plots(B_plots, ncol = 3, guides = "collect") +
  plot_annotation(
    title = "B cell markers expression on the UMAP projection"
  ) &
  theme(
    legend.position = "right",
    legend.title    = element_text(face = "bold"),
    legend.text     = element_text(size = 8)
  )

B_feature_grid

ggsave(file.path(main_dir, "Results/DE_Plots/FeaturePlots/B_markers_UMAP_grid.png"),
       B_feature_grid,
  width  = 16,
  height = 10,
  dpi    = 300, device = "png"
)
```

```{r Violin plots for the unsupervised results}

Custom_VLN(obj = Seurat_obj, 
           features = c("ALDOB"),
           group_by = "Unsupervised_3_clusters",
           ncol = 1)


plots_Uns_3 <- Plot_top_markers_VLN(obj = Seurat_obj,
                                  markers_tbl = markers_unsupervised,
                                  group_by = "Unsupervised_3_clusters",
                                  out_dir = file.path(main_dir, "Results/DE_Plots/VLN_Plots/Unsupervised_3"),
                                  top_n = 3,
                                  prefix = "Unsupervised_3")

B_pan <- c("MS4A1", "CD79A", "CD79B", "CD19", "CD74", "CD22")
intersect(B_pan, Features(Seurat_obj))
gene_tag <- paste(B_pan, collapse = "_")

p1 <- Plot_manual_markers_VLN(obj = Seurat_obj,
                              group_by = "Unsupervised_3_clusters",
                              cell_type = "B.cell",
                              markers = B_pan,
                              save_path = file.path(main_dir, "Results/DE_Plots/VLN_Plots/Custom_Unsupervised_3",
                              paste0("B_cell_", gene_tag, ".png")),
                              pt_size = 0)

```

```{r Plot all differentially expressed markers for a cluster group}

## Unsupervised model ##
Plot_all_sig_genes_VLN(obj = Seurat_obj, 
                       markers_tbl = markers_unsupervised, 
                       group_by = "Unsupervised_3_clusters", 
                       annotation_name = "Unsupervised_3", 
                       out_dir_base = file.path(main_dir, "Results/DE_Plots/VLN_Plots"), 
                       p_adj_cutoff = 0.05, 
                       avg_log2FC_cutoff = 1.5,
                       pt_size = 0, 
                       overwrite = TRUE)

## Reference based - Kidney HCA ##
Plot_all_sig_genes_VLN(obj = Seurat_obj, 
                       markers_tbl = markers_Kidney_HCA, 
                       group_by = "Kidney_HCA_clusters", 
                       annotation_name = "Kidney_HCA", 
                       out_dir_base = file.path(main_dir, "Results/DE_Plots/VLN_Plots"), 
                       p_adj_cutoff = 0.05, 
                       avg_log2FC_cutoff = 1.5,
                       pt_size = 0, 
                       overwrite = TRUE)

```

Create the dataframe for the manual annotation markers based on this publication: https://doi.org/10.1146/annurev-physiol-052521-121841

```{r Vasculature Markers}

Endothilial <- c("NRP1", "CDH5", "ELN")
Glomerular_Endothilium <- c("PLAT", "EMCN", "TSAPN7", "MAPT", 
                            "KDR", "SMAD6", "EHD3", "LPL", "FLT1", 
                            "FBLN2", "MGP", "TRPV4", "BMX")
Capillaries <- c("KDR", "SMAD6", "EHD3", "LPL", "FLT1")
Arterioles <- c("FBLN2", "MGP", "TRPV4", "BMX",
                "SOX17", "CXCL12", "GJA5")
Vas_afferens <- c("EDN1", "FBLN5", "CLDN5", "EFNB2")
Vas_efferens <- c("KLF4", "CRYAB", "GAS6", "PODX")
Peritubular_capillaries <- c("PLVAP")
Venules <- c("PIVAP", "BGN", "CD9", "NR2F2")
Ascending_vasa_recta <- c("FXYD2", "FXYD6", "IGFOP7")
Descending_vasa_recta <- c("SIC14A1", "AQP1", "S100A4")

Vasc_markers_list <- list(Endothilial = Endothilial, 
                          Glomerular_Endothilium = Glomerular_Endothilium, 
                          Capillaries = Capillaries, 
                          Arterioles = Arterioles, 
                          Vas_afferens = Vas_afferens, 
                          Vas_efferens = Vas_efferens, 
                          Peritubular_capillaries = Peritubular_capillaries, 
                          Venules = Venules, 
                          Ascending_vasa_recta = Ascending_vasa_recta, 
                          Descending_vasa_recta = Descending_vasa_recta)

Vasc_markers_df <- enframe(Vasc_markers_list, name = "Subtype", value = "Gene") %>% 
  tidyr::unnest_longer(Gene) %>% 
  mutate(Type = "Vasculature", .before = Subtype)

```

```{r Proximal tubule Markers}

Pan_PT <- c("SLC34A1", "LRP2", "HXYD2", "HRSP12", "ACSM1", 
            "ACSM2", "CPT1A", "ACOX3", "SLC26A6", "SLC9A3", 
            "GLUD1", "PCK1", "AQP8", "HNF4A", "PPARA")
Progenitors_PT <- c("NOTCH2", "LGR4")
Proximal_Tubule_straight <- c("ATP11A", "SLC13A3", "SLC16A9", "SLC27A2", 
                              "SLC7A13", "SLC22A6", "SLC1A1")
Proximal_Tubule_convoluted <- c("ATP5A2", "SLC5A12", "ADRA1A", "SLC6A19", 
                                "SLC7A8", "SLC7A9")
Injured_PT <- c("HAVCR1", "KRT20", "HSPA1A", "VCAM1", 
                "DCDC2A", "SEMA5A")

PT_markers_list <- list(Pan_PT = Pan_PT,
                        Progenitors_PT = Progenitors_PT,
                        Injured_PT = Injured_PT,
                        Proximal_Tubule_straight = Proximal_Tubule_straight,
                        Proximal_Tubule_convoluted = Proximal_Tubule_convoluted)

PT_markers_df <- enframe(PT_markers_list, name = "Subtype", value = "Gene") %>% 
  tidyr::unnest_longer(Gene) %>% 
  mutate(Type = "Proximal_Tubule", .before = Subtype)

```

```{r Mesangium (smooth muscle cells and juxtaglomerular cells) Markers}

Mesangial <- c("Serpine2", "FHL2", "DES", "PRKCA", 
               "ART3", "NT5E", "PDGFRB")
Pericytes <- c("VIM", "TAGLN", "MYH11", "PDGFRB")
SMC <- c("TAGLN", "MYH11", "ACTA2", 
         "GATA3", "RERGL", "MAP3K7D") 
JG <- c("REN1", "AKR1B7", "RGS5")

Mesangium_markers_list <- list(Mesangial = Mesangial,
                               Pericytes = Pericytes,
                               SMC = SMC,
                               JG = JG)

Mesangium_markers_df <- enframe(Mesangium_markers_list, name = "Subtype", value = "Gene") %>% 
  tidyr::unnest_longer(Gene) %>% 
  mutate(Type = "Mesangium", .before = Subtype)

```

```{r Podocyte Markers}

Adult_Podocytes <- c("NPHS1", "NPHS2", "SYNPO", "CDKN1C", "WT1")
Progenitor_Podocytes <- c("WT1", "FOXC2", "MAFB", "EFNB2", "FOXL1")

Podocyte_markers_list <- list(Adult_Podocytes = Adult_Podocytes,
                               Progenitor_Podocytes = Progenitor_Podocytes)

Podocyte_markers_df <- enframe(Podocyte_markers_list, name = "Subtype", value = "Gene") %>% 
  tidyr::unnest_longer(Gene) %>% 
  mutate(Type = "Podocyte", .before = Subtype)

```

```{r Loop of Henly and Macula Densa Markers}

Descending_thin_LOH <- c("FST", "AQP1", "SLC14A2", "BST1", "EPHA7", 
                         "CRYAB", "TSHZ2", "CALD1", "LYPD2")
Assending_thin_LOH <- c("EPHA7", "MX2", "CLCNKA")
Assending_thick_LOH <- c("SLC12A1", "UMOD", "TMEM207", "FOXQ1", 
                         "CLDN10", "PTGER3", "KCNJ1", "ENOX1", 
                         "THSD4", "MT2", "SLC5A3")

LOH_markers_list <- list(Descending_thin_LOH = Descending_thin_LOH,
                         Assending_thin_LOH = Assending_thin_LOH,
                         Assending_thick_LOH = Assending_thick_LOH)

LOH_markers_df <- enframe(LOH_markers_list, name = "Subtype", value = "Gene") %>% 
  tidyr::unnest_longer(Gene) %>% 
  mutate(Type = "LOH", .before = Subtype)

Macula_Densa <- c("ENOX1", "THSD4", "NAS1", "AVPR1A")

Macula_Densa_markers_list <- list(Macula_Densa = Macula_Densa)

Macula_Densa_markers_df <- enframe(Macula_Densa_markers_list, name = "Subtype", value = "Gene") %>% 
  tidyr::unnest_longer(Gene) %>% 
  mutate(Type = "MD", .before = Subtype)

```

```{r Distal convoluted and connecting tubule}

DCT1 <- c("PVALB", "SLC12A3", "TRPM7", "WNK1", "WNK4", 
          "STK39", "CALB1", "SLC8A1", "EGF", "TRPM6", "FXYD2",
          "CNNM2", "ATP1A1", "ATP1A2", "ATP1A3", "ATP1A4")
DCT2 <- c("SLC12A3", "TRPM7", "WNK1", "WNK4", "KLHL3", 
          "STK39", "CALB1", "SLC8A1", "EGF", "TRPM6", 
          "CNNM2", "ATP1A1", "ATP1A2", "ATP1A3", "ATP1A4", 
          "FXYD2", "KLK1", "TRPV5", "S100G", "ATP2B1", 
          "ATP2B4", "SCNN1B", "SCNN1G", "KCNE1")

DCT_markers_list <- list(DCT1 = DCT1,
                         DCT2 = DCT2)

DCT_markers_df <- enframe(DCT_markers_list, name = "Subtype", value = "Gene") %>% 
  tidyr::unnest_longer(Gene) %>% 
  mutate(Type = "DCT", .before = Subtype)

CNT <- c("CALB1", "SLC8A1", "EGF", "KLK1", 
         "TRPM6", "S100G", "ATP2B1", "SCNN1B", 
         "SCNN1G", "KCNE1")

CNT_markers_list <- list(CNT = CNT)

CNT_markers_df <- enframe(CNT_markers_list, name = "Subtype", value = "Gene") %>% 
  tidyr::unnest_longer(Gene) %>% 
  mutate(Type = "CNT", .before = Subtype)

```

```{r Collecting duct markers}

CD_Principal <- c("SCNN1B", "SCNN1G", "AQP2", "AVPR2", "HSD11B2", 
                  "RHBG", "ELF5", "FXYD4", "AQP3", "APELA", 
                  "KCNE1", "NPNT", "KCNJ10")
CD_Intercalated_Pan <- c("TCFCP2L1", "FOXI", "ATP6V1G3", 
                         "ATP6V0D2", "INSR", "ATP6V1b1")
CD_Intercalated_A <- c("ATP4A", "SLC4A1", "AQP6", 
                       "KIT", "ADGRF5", "MME")
CD_Intercalated_B <- c("SLC26A4", "HMX2", "SPINK8")
CD_Transitional <- c("AQP2", "HSD11B2", "RHBG", "ATP6V1G3", "ATP6V0D2", 
                     "INSR", "ATP6V1B1", "PARM1", "SEC23B")

CD_markers_list <- list(CD_Principal = CD_Principal,
                        CD_Intercalated_Pan = CD_Intercalated_Pan, 
                        CD_Intercalated_A = CD_Intercalated_A, 
                        CD_Intercalated_B = CD_Intercalated_B, 
                        CD_Transitional = CD_Transitional)

CD_markers_df <- enframe(CD_markers_list, name = "Subtype", value = "Gene") %>% 
  tidyr::unnest_longer(Gene) %>% 
  mutate(Type = "CD", .before = Subtype)

```

```{r Immune cell markers}

Macrophages <- c("C1QA", "C1QB", "ITGAM", "APOE", "C1QC", 
                  "CD74", "CTSS", "FCER1G", "AIF1", "MS4A7")
Neutrophils <- c("S100A8", "S100A9", "LYZ2", "PLAC8", "IFITM3", 
                 "CEBPB", "TYROBP", "LST1", "FCER1G", "HP")
Basophils <- c("IFITM1", "HDC", "MCMPT8", "FCER1A", "CRSP3", 
               "MS4A2", "CYP11A1", "CD200R3", "IL6", "IL4")
Dendritic_11B_pos <- c("CD74", "CD209A", "WFDC17", "MGL2", "CCL6", 
                       "CCL9", "CTSS", "ALOX5AP", "IFITM3", "TYROBP")
Dendritic_11B_neg <- c("IRF8", "NAAA", "LPBD1", "CBFA2T3", "BASP1", 
                       "RNASE6", "WDFY3", "SEPT3", "PPM1M", "RAB7B")
Dendritic_plasmocytoid <- c("LY6D", "SIGLECH", "COX6A2", "RNASE6", "SELL", 
                            "CCR9", "RUNX2", "CD209D", "BCL11A", "LAIR1")
B_Cells <- c("CD79A", "CD79B", "MS4A1", "LY6D", "EBF1", 
             "CD22", "CD19", "FCMR", "SIGLECG", "FCRL1")
T_Cells <- c("CXCR6", "CD247", "NKG7")
CD4_T_Cells <- c("LEF1", "MS4A4B", "IL7R", "CCR7", "KLF2", 
                 "TCF7", "DAPL1", "SATB1", "CD3D")
CD8_Effector <- c("CCL5", "NKG7", "IL7R", "CCR7", "KLF2", 
                  "TCF7", "DAPL1", "SATB1", "CD3D")
T_regulatory <- c("TNFRSF4", "CAPG", "IKZF2", "IZUMO1R", "IFI27L2A", 
                  "S100A4", "RGS1", "CD3G", "LTB", "TNFRSF18")
T_NK_Cells <- c("LY6C2", "CXCR6", "GIMAP3", "TMSB10", "CD3G", 
                "GIMAP4", "CTSW", "NKG7", "HCST", "LTB")
NK_Cells <- c("GZMA", "NKG7", "CD7", "CCL5", "XCL1", 
              "KLRD1", "KLRK1", "NCR1", "KLRE1", "IL2RB")

Immune_markers_list <- list(Macrophages = Macrophages,
                            Neutrophils = Neutrophils, 
                            Basophils = Basophils, 
                            Dendritic_11B_pos = Dendritic_11B_pos, 
                            Dendritic_11B_neg = Dendritic_11B_neg, 
                            Dendritic_plasmocytoid = Dendritic_plasmocytoid, 
                            B_Cells = B_Cells, 
                            T_Cells = T_Cells, 
                            CD4_T_Cells = CD4_T_Cells, 
                            CD8_Effector = CD8_Effector, 
                            T_regulatory = T_regulatory, 
                            T_NK_Cells = T_NK_Cells, 
                            NK_Cells = NK_Cells)

Immune_markers_df <- enframe(Immune_markers_list, name = "Subtype", value = "Gene") %>% 
  tidyr::unnest_longer(Gene) %>% 
  mutate(Type = "Immune", .before = Subtype)

```

```{r Bind the marker dataframes}

Kidney_Markers_total <- bind_rows(Vasc_markers_df, PT_markers_df, 
                                  Mesangium_markers_df, Podocyte_markers_df, 
                                  Macula_Densa_markers_df, CNT_markers_df, 
                                  CD_markers_df, Immune_markers_df)

```
